fix: adding security to server actions

+++ app/_actions/deletePost.ts
@@ -5,6 +5,17 @@ import { revalidatePath } from "next/cache";
 
 export async function deletePost(slug: string) {
     const supabase = await createClient();
+
+    const {
+        data: { user },
+        error: authError,
+    } = await supabase.auth.getUser();
+    if (authError || !user) {
+        return {
+            error: "Unauthorized: You must be logged in to update posts.",
+        };
+    }
+
     const { error } = await supabase
         .from("blog_posts")
         .delete()
